0910 cho - Order Update

+++ src/main/java/com/project/erpre/model/Customer.java
@@ -20,7 +20,7 @@ public class Customer {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "customer_no")
-    private Long customerNo;
+    private Integer customerNo;
 
     @Column(name = "customer_name", length = 100, nullable = false)
     private String customerName;
